Framework update (#48)
* f719340 Reinstate TZ
* 8a46629 Support long longs in ArduinoJson
* f39229d Speed uploads
* cda685f Remove deprecated declaration
* 3316a14 Remove millis from uptime
* 6926790  Remove potential buffer overflow
* fd76370 Use updated lib
* 121fafe Remove unused method
* 4778c1e Get rid of issue with const

+++ src/thingSpeakWork.h
@@ -46,7 +46,7 @@ int createChannel(
     String &channelJSON,
     const char *api_key,
     const char *name,
-    const unsigned long fieldCount,
+    const int fieldCount,
     const char *field[],
     const char *description = "",
     const bool public_flag = false,
